+++ src/config/options.ts
@@ -30,7 +30,6 @@ export const redisOptions = {
 }
 
 export const psqlOptions = {
-    type: 'postgres',
     host: getString('POSTGRES_DB_HOST'),
     port: getNumber('POSTGRES_DB_PORT'),
     username: getString('POSTGRES_DB_USERNAME'),
